0205-7
型を修正

+++ ex02/PmergeMe.hpp
@@ -18,8 +18,8 @@ class PmergeMe
 		static void displaySequence(std::string const&, std::vector<int> const&);
 		static void displayTime(std::string const&, clock_t, clock_t);
 
-		static void mergeInsertionSort(std::vector<int> &, int);
-		static void mergeInsertionSortHelper(std::vector<int> &, std::vector<int> &, int, int, int);
-		static void insertionSortRange(std::vector<int> &, int, int);
-		static void merge(std::vector<int> &, std::vector<int> &, int, int, int);
+		static void mergeInsertionSort(std::vector<int> &, size_t);
+		static void mergeInsertionSortHelper(std::vector<int> &, std::vector<int> &, size_t, size_t, size_t);
+		static void insertionSortRange(std::vector<int> &, size_t, size_t);
+		static void merge(std::vector<int> &, std::vector<int> &, size_t, size_t, size_t);
 };
